Move non-int64 sparse tests to test_sparray; dedup and fold
Addresses review on the sparray scipy-parity split: the int64 test
module had accumulated general scipy-parity tests unrelated to int64
indices, plus duplicates of coverage that already existed elsewhere.

- Move the pure scipy-parity classes (negative-shape rejection,
  cross-format comparison, 2-D setdiag rejection, indptr[0]==0 check,
  complex sign) from test_sparse_int64_indices.py to test_sparray.py.
- Drop duplicates (in-place scalar specials, bool/int promotion,
  power-zero, csrilu02 smoke test) and fold their unique cases into the
  existing test_sparray.py tests; use `is` instead of id() for the
  identity checks.
- Clarify the "tight buffers" comment in _compressed.py and tag the
  complex-sign workaround in _data.py with TODO(eriknw): pointing
  at gh-10034.
