+++ tests/cupyx_tests/scipy_tests/sparse_tests/test_sparse_int64_indices.py
@@ -3550,118 +3550,6 @@ class TestInt64TransposeCanonical:
         assert not hasattr(c, '_has_canonical_format')
 
 
-class TestInplaceScalarSpecials:
-
-    @pytest.mark.parametrize('cls_name', ['csr_matrix', 'csc_matrix',
-                                          'coo_matrix'])
-    def test_imul_scalar_preserves_identity(self, cls_name):
-        cls = getattr(sparse, cls_name)
-        A = cls(cupy.array([[1.0, 2.0], [3.0, 4.0]]))
-        old = A
-        old_data_id = id(A.data)
-        A *= 2
-        assert A is old
-        assert id(A.data) == old_data_id
-        cupy.testing.assert_array_equal(
-            A.toarray(), cupy.array([[2.0, 4.0], [6.0, 8.0]]))
-
-    @pytest.mark.parametrize('cls_name', ['csr_matrix', 'csc_matrix',
-                                          'coo_matrix'])
-    def test_itruediv_scalar_preserves_identity(self, cls_name):
-        cls = getattr(sparse, cls_name)
-        A = cls(cupy.array([[2.0, 4.0], [6.0, 8.0]]))
-        old = A
-        A /= 2
-        assert A is old
-        cupy.testing.assert_array_equal(
-            A.toarray(), cupy.array([[1.0, 2.0], [3.0, 4.0]]))
-
-    def test_imul_dia(self):
-        # DIA only accepts the (data, offsets) constructor.
-        data = cupy.array([[1.0, 2.0, 3.0]])
-        offsets = cupy.array([0])
-        A = sparse.dia_matrix((data, offsets), shape=(3, 3))
-        old = A
-        A *= 2
-        assert A is old
-        cupy.testing.assert_array_equal(
-            A.toarray(),
-            cupy.array([[2.0, 0.0, 0.0],
-                        [0.0, 4.0, 0.0],
-                        [0.0, 0.0, 6.0]]))
-
-    def test_imul_non_scalar_falls_back(self):
-        # For non-scalar operands, ``__imul__`` returns NotImplemented
-        # so Python rebinds via ``A = A * other``.  Identity is NOT
-        # preserved -- this is the same as scipy's behavior.
-        A = sparse.csr_matrix(cupy.array([[1.0, 0.0], [0.0, 1.0]]))
-        B = sparse.csr_matrix(cupy.array([[2.0, 0.0], [0.0, 3.0]]))
-        old = A
-        A *= B
-        # element-wise multiply is delegated to ``A.multiply(B)``;
-        # the result is a new object.
-        assert A is not old
-        cupy.testing.assert_array_equal(
-            A.toarray(), cupy.array([[2.0, 0.0], [0.0, 3.0]]))
-
-    def test_imul_scalar_zero(self):
-        # ``A *= 0`` zeroes out the data buffer in place but leaves
-        # the structure (indices/indptr).  Matches scipy.
-        A = sparse.csr_matrix(cupy.array([[1.0, 2.0], [3.0, 4.0]]))
-        old = A
-        original_indices = A.indices.copy()
-        A *= 0
-        assert A is old
-        cupy.testing.assert_array_equal(
-            A.toarray(), cupy.zeros((2, 2)))
-        # Structure is preserved: explicit zeros remain stored.
-        cupy.testing.assert_array_equal(A.indices, original_indices)
-
-
-class TestNegativeShapeRejected:
-    # Error message matches scipy: "'shape' elements cannot be negative".
-
-    @pytest.mark.parametrize('cls_name', ['csr_matrix', 'csc_matrix',
-                                          'coo_matrix', 'csr_array',
-                                          'csc_array', 'coo_array',
-                                          'dia_matrix', 'dia_array'])
-    @pytest.mark.parametrize('shape', [(10, -5), (-5, 10), (-5, -5)])
-    def test_negative_shape_raises(self, cls_name, shape):
-        cls = getattr(sparse, cls_name)
-        if cls_name.startswith('dia'):
-            # DIA only accepts (data, offsets) + shape= kwarg.
-            with pytest.raises(
-                    ValueError,
-                    match=r"'shape' elements cannot be negative"):
-                cls((cupy.array([[1.0]]), cupy.array([0])), shape=shape)
-        else:
-            with pytest.raises(
-                    ValueError,
-                    match=r"'shape' elements cannot be negative"):
-                cls(shape)
-
-    @pytest.mark.parametrize('cls_name', ['csr_matrix', 'csc_matrix',
-                                          'coo_matrix'])
-    def test_negative_shape_kwarg_raises(self, cls_name):
-        # Same check via the shape= keyword on a 3-tuple constructor.
-        cls = getattr(sparse, cls_name)
-        data = cupy.array([1.0])
-        if cls_name == 'coo_matrix':
-            row = cupy.array([0], dtype='i')
-            col = cupy.array([0], dtype='i')
-            with pytest.raises(
-                    ValueError,
-                    match=r"'shape' elements cannot be negative"):
-                cls((data, (row, col)), shape=(10, -5))
-        else:
-            indices = cupy.array([0], dtype='i')
-            indptr = cupy.array([0, 1, 1, 1], dtype='i')
-            with pytest.raises(
-                    ValueError,
-                    match=r"'shape' elements cannot be negative"):
-                cls((data, indices, indptr), shape=(10, -5))
-
-
 class TestCanonicalGetterPropagatesSorted:
     """``has_canonical_format=True`` implies sorted indices, so the
     canonical getter must also cache ``_has_sorted_indices=True`` --
@@ -4211,46 +4099,6 @@ class TestAsindicesOutOfBoundsRejected:
             a[[2**40]]
 
 
-class TestBoolIntScalarPromotion:
-    # cuSPARSE accepts only '?fdFD' for data; promote bool/int to
-    # float64 so the result is usable downstream.
-
-    def test_bool_imul_int(self):
-        a = sparse.csr_matrix(cupy.array([[True, False], [False, True]]))
-        old = a
-        a *= 2
-        assert a is old
-        assert a.dtype == cupy.float64
-        cupy.testing.assert_array_equal(a.data, cupy.array([2.0, 2.0]))
-
-    def test_bool_idiv_int(self):
-        a = sparse.csr_matrix(cupy.array([[True, False], [False, True]]))
-        old = a
-        a /= 2
-        assert a is old
-        assert a.dtype == cupy.float64
-        cupy.testing.assert_array_equal(a.data, cupy.array([0.5, 0.5]))
-
-    def test_bool_truediv_int(self):
-        # Non-in-place truediv goes through CSR's ``__truediv__``.
-        a = sparse.csr_matrix(cupy.array([[True, False], [False, True]]))
-        b = a / 2
-        assert b.dtype == cupy.float64
-        cupy.testing.assert_array_equal(b.data, cupy.array([0.5, 0.5]))
-
-    def test_float32_truediv_keeps_float64(self):
-        # Regression: float32 / 2.0 -> float64 (matches scipy's rule).
-        a = sparse.csr_matrix(
-            cupy.array([[1.0, 2.0]], dtype=cupy.float32))
-        b = a / 2.0
-        assert b.dtype == cupy.float64
-
-    def test_complex_truediv_unchanged(self):
-        a = sparse.csr_matrix(cupy.array([[1+2j, 3+4j]]))
-        b = a / 2
-        assert b.dtype == cupy.complex128
-
-
 class TestSparseTypingAliases:
 
     def test_csr_array_typing(self):
@@ -4262,161 +4110,3 @@ class TestSparseTypingAliases:
         import types
         alias = sparse.csr_matrix[float, int]
         assert isinstance(alias, types.GenericAlias)
-
-
-class TestPowerZeroRaises:
-    """``A.power(0)`` would densify the matrix (every implicit zero
-    becomes ``0**0 == 1``); raise NotImplementedError to match scipy
-    instead of returning a sparse-but-mathematically-wrong result.
-    """
-
-    @pytest.mark.parametrize('cls_name', ['csr_matrix', 'csr_array',
-                                          'csc_matrix', 'csc_array',
-                                          'coo_matrix', 'coo_array'])
-    def test_method_raises(self, cls_name):
-        cls = getattr(sparse, cls_name)
-        a = cls(cupy.array([[1.0, 0, 2.0], [0, 3.0, 0]]))
-        with pytest.raises(NotImplementedError, match='zero power'):
-            a.power(0)
-
-    def test_array_op_raises(self):
-        # sparray ``A ** 0`` -- must catch even though base also has a
-        # check (both paths route through ``power`` now).
-        a = sparse.csr_array(cupy.array([[1.0, 0, 2.0]]))
-        with pytest.raises(NotImplementedError, match='zero power'):
-            a ** 0
-
-    def test_array_pow_array_exponent(self):
-        # The non-scalar check must run before any ``other == 0``
-        # evaluation, otherwise ``if other == 0`` raises "truth value
-        # ambiguous" on an array exponent.
-        a = sparse.csr_array(cupy.array([[1.0, 0, 2.0]]))
-        with pytest.raises(NotImplementedError, match='not scalar'):
-            a ** cupy.array([2, 3, 4])
-
-    def test_method_nonzero_works(self):
-        a = sparse.csr_matrix(cupy.array([[2.0, 4.0]]))
-        b = a.power(2)
-        cupy.testing.assert_array_equal(b.data, cupy.array([4.0, 16.0]))
-
-
-class TestComparisonCrossFormat:
-    """``_comparison`` and ``_maximum_minimum`` accept any sparse
-    operand by routing through ``other.tocsr()`` (matches
-    ``_add_sparse`` / ``multiply``), so ``csr.maximum(coo)``,
-    ``csr == csc`` etc. don't bottom out in NotImplementedError.
-    """
-
-    def test_maximum_csr_coo(self):
-        a = sparse.csr_matrix(cupy.array([[1.0, 2.0], [3.0, 4.0]]))
-        b = sparse.coo_matrix(cupy.array([[5.0, 1.0], [2.0, 8.0]]))
-        c = a.maximum(b)
-        cupy.testing.assert_array_equal(
-            c.toarray(),
-            cupy.array([[5.0, 2.0], [3.0, 8.0]]))
-
-    def test_minimum_csr_csc(self):
-        a = sparse.csr_matrix(cupy.array([[5.0, 6.0], [7.0, 8.0]]))
-        b = sparse.csc_matrix(cupy.array([[1.0, 2.0], [3.0, 4.0]]))
-        c = a.minimum(b)
-        cupy.testing.assert_array_equal(
-            c.toarray(),
-            cupy.array([[1.0, 2.0], [3.0, 4.0]]))
-
-    def test_eq_csr_coo(self):
-        import warnings
-        a = sparse.csr_matrix(cupy.array([[1.0, 2.0]]))
-        b = sparse.coo_matrix(cupy.array([[1.0, 0.0]]))
-        with warnings.catch_warnings():
-            warnings.simplefilter('ignore', sparse.SparseEfficiencyWarning)
-            c = a == b
-        assert sparse.issparse(c)
-
-    def test_lt_csr_dia(self):
-        a = sparse.csr_matrix(cupy.array([[1.0, 2.0], [3.0, 4.0]]))
-        b = sparse.dia_matrix(
-            (cupy.array([[5.0, 6.0]]), cupy.array([0])),
-            shape=(2, 2))
-        c = a < b
-        assert sparse.issparse(c)
-
-
-class TestSetdiag2DRejected:
-
-    def test_csr_setdiag_2d_raises(self):
-        a = sparse.csr_matrix(cupy.zeros((3, 3)))
-        with pytest.raises(ValueError, match='must be 0-d or 1-d'):
-            a.setdiag(cupy.array([[1.0, 2.0], [3.0, 4.0]]))
-
-    def test_coo_setdiag_2d_raises(self):
-        a = sparse.coo_matrix(cupy.zeros((3, 3)))
-        with pytest.raises(ValueError, match='must be 0-d or 1-d'):
-            a.setdiag(cupy.array([[1.0, 2.0], [3.0, 4.0]]))
-
-
-class TestPublicCtorIndptrZeroCheck:
-
-    def test_csr_indptr_must_start_at_zero(self):
-        with pytest.raises(ValueError, match='start with 0'):
-            sparse.csr_matrix(
-                (cupy.array([1.0, 2.0]),
-                 cupy.array([0, 1], dtype='i'),
-                 cupy.array([5, 6, 7], dtype='i')),
-                shape=(2, 2))
-
-    def test_csc_indptr_must_start_at_zero(self):
-        with pytest.raises(ValueError, match='start with 0'):
-            sparse.csc_matrix(
-                (cupy.array([1.0, 2.0]),
-                 cupy.array([0, 1], dtype='i'),
-                 cupy.array([5, 6, 7], dtype='i')),
-                shape=(2, 2))
-
-
-class TestComplexSignNumpy2x:
-    """scipy 1.16+ uses numpy 2.x ``sign`` semantics for complex
-    (``z / abs(z)``).  ``cupy.sign`` follows the same rule but
-    returns ``nan+nanj`` for ``0+0j`` (literal ``0/0``); mask
-    explicit zeros to keep stored ``0+0j`` round-tripping cleanly.
-    """
-
-    def test_complex_sign_unit_vector(self):
-        a = sparse.csr_matrix(cupy.array([[1+2j]]))
-        b = a.sign()
-        # (1+2j) / |1+2j| = (1+2j) / sqrt(5) ≈ 0.447 + 0.894j
-        cupy.testing.assert_allclose(
-            b.data,
-            cupy.array([1+2j]) / cupy.abs(cupy.array([1+2j])))
-
-    def test_complex_sign_zero_masked(self):
-        # Stored ``0+0j`` should round-trip to ``0+0j``, not ``nan+nanj``.
-        a = sparse.csr_matrix._from_parts(
-            cupy.array([0+0j, 1+1j]),
-            cupy.array([0, 1], dtype='i'),
-            cupy.array([0, 1, 2], dtype='i'),
-            (2, 2))
-        b = a.sign()
-        # First entry is 0+0j (masked), second is unit vector.
-        assert b.data[0] == 0+0j
-        cupy.testing.assert_allclose(
-            cupy.abs(b.data[1]).item(), 1.0, atol=1e-6)
-
-    def test_real_sign_unchanged(self):
-        a = sparse.csr_matrix(cupy.array([[-2.0, 0, 3.0]]))
-        b = a.sign()
-        cupy.testing.assert_array_equal(b.data, cupy.array([-1.0, 1.0]))
-
-
-class TestCsrilu02NoLeak:
-
-    def test_csrilu02_runs(self):
-        # Smoke test: the descriptor leak isn't observable at the
-        # Python level; this just verifies the try/finally wrapping
-        # didn't break the happy path.
-        from cupyx import cusparse
-        a = sparse.csr_matrix(cupy.array(
-            [[2.0, 0.1, 0, 0], [0.1, 2.0, 0.1, 0],
-             [0, 0.1, 2.0, 0.1], [0, 0, 0.1, 2.0]],
-            dtype=cupy.float64))
-        a.sum_duplicates()
-        cusparse.csrilu02(a)
